fix(build): fix eslint
