+++ src/features/company-profile/components/CompanyHeader.tsx
@@ -1,6 +1,7 @@
-"use client";
+'use client';
 
 import { Button } from '@/components/ui/button';
+import Image from 'next/image';
 import { Mail, Phone, Building, Edit } from 'lucide-react';
 import type { Company } from '@/types/company';
 
@@ -13,63 +14,64 @@ export default function CompanyHeader({ company, viewType }: CompanyHeaderProps)
   return (
     <div className="relative">
       {/* Banner */}
-      <div 
+      <div
         className="h-85 bg-contain bg-center"
         style={{
-          backgroundImage: company.bannerUrl ? `url(${company.bannerUrl})` : undefined
+          backgroundImage: company.bannerUrl ? `url(${company.bannerUrl})` : undefined,
         }}
       />
-      
+
       {/* Company Info Card */}
-      <div className="relative -mt-24 mx-auto max-w-6xl px-6">
-        <div className="bg-very-dark-gray border border-zinc-700 rounded-xl p-8 shadow-xl">
-          <div className="flex flex-col md:flex-row items-start gap-6">
+      <div className="relative mx-auto -mt-24 max-w-6xl px-6">
+        <div className="bg-very-dark-gray rounded-xl border border-zinc-700 p-8 shadow-xl">
+          <div className="flex flex-col items-start gap-6 md:flex-row">
             {/* Company Logo */}
             <div className="flex-shrink-0">
-              <div className="w-24 h-24 bg-component rounded-xl flex items-center justify-center border border-zinc-600 overflow-hidden">
+              <div className="bg-component flex h-24 w-24 items-center justify-center overflow-hidden rounded-xl border border-zinc-600">
                 {company.logoUrl ? (
-                  <img 
-                    src={company.logoUrl} 
-                    alt={`${company.name} logo`}
-                    className="w-full h-full object-cover"
-                  />
+                  <div className="relative h-full w-full">
+                    <Image
+                      src={company.logoUrl}
+                      alt={`${company.name} logo`}
+                      fill
+                      sizes="96px"
+                      className="object-cover"
+                      unoptimized
+                    />
+                  </div>
                 ) : (
-                  <Building className="w-12 h-12 text-primary-green" />
+                  <Building className="text-primary-green h-12 w-12" />
                 )}
               </div>
             </div>
-            
+
             {/* Company Details */}
-            <div className="flex-1 min-w-0">
-              <div className="flex flex-col md:flex-row md:items-start md:justify-between gap-4">
+            <div className="min-w-0 flex-1">
+              <div className="flex flex-col gap-4 md:flex-row md:items-start md:justify-between">
                 <div>
-                  <h1 className="text-3xl font-bold text-white mb-2">
-                    {company.name}
-                  </h1>
+                  <h1 className="mb-2 text-3xl font-bold text-white">{company.name}</h1>
                   <p className="text-lighter-gray-text mb-4">
                     {company.industry} | {company.employeeCount} | {company.location}
                   </p>
-                  
+
                   {/* Contact Info */}
-                  <div className="flex flex-col sm:flex-row gap-4 text-sm text-lighter-gray-text">
+                  <div className="text-lighter-gray-text flex flex-col gap-4 text-sm sm:flex-row">
                     <div className="flex items-center gap-2">
-                      <Mail className="w-4 h-4" />
+                      <Mail className="h-4 w-4" />
                       <span>{company.email}</span>
                     </div>
                     <div className="flex items-center gap-2">
-                      <Phone className="w-4 h-4" />
+                      <Phone className="h-4 w-4" />
                       <span>{company.phone}</span>
                     </div>
                   </div>
                 </div>
-                
+
                 {/* Action Button */}
                 <div className="flex-shrink-0">
                   {viewType === 'company' && (
-                    <Button 
-                      className="bg-[#595256] hover:bg-gray-cancel text-white px-6 py-2 rounded-md"
-                    >
-                      <Edit className="w-4 h-4" />
+                    <Button className="hover:bg-gray-cancel rounded-md bg-[#595256] px-6 py-2 text-white">
+                      <Edit className="h-4 w-4" />
                       Edit Profile
                     </Button>
                   )}
