+++ src/features/landing/components/GoogleLogin.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import Image from 'next/image';
 
 type GoogleLoginProps = Readonly<{
   onClick?: () => void;
@@ -16,11 +17,16 @@ export default function GoogleLogin({
       onClick={onClick}
       className={`flex cursor-pointer items-center gap-2 rounded-lg border border-gray-300 bg-white px-4 py-2 shadow-sm transition hover:bg-gray-200 ${className}`}
     >
-      <img
-        src="https://www.svgrepo.com/show/475656/google-color.svg"
-        alt="Google logo"
-        className="h-5 w-5"
-      />
+      <div className="relative h-5 w-5">
+        <Image
+          src="https://www.svgrepo.com/show/475656/google-color.svg"
+          alt="Google logo"
+          width={20}
+          height={20}
+          unoptimized
+          className="object-contain"
+        />
+      </div>
       <span className="font-medium text-gray-700">{label}</span>
     </button>
   );
